@@ -20,14 +20,14 @@
   },
   "dependencies": {
     "@kentcdodds/react-workshop-app": "file:../..",
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^11.2.6",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.1",
     "chalk": "^4.1.0",
     "codegen.macro": "^4.1.0",
     "is-ci-cli": "^2.2.0",
     "jest-environment-jsdom-sixteen": "^1.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "typescript": "^4.2.3"
   },
   "devDependencies": {
