@@ -64,10 +64,10 @@
     "@emotion/core": "^10.0.35",
     "@reach/tabs": "^0.14.0",
     "@reach/tooltip": "^0.14.0",
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^11.2.6",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.1",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "chalk": "^4.1.0",
     "codegen.macro": "^4.1.0",
     "cross-env": "^7.0.3",
@@ -104,8 +104,8 @@
     "postcss": "^8.2.9",
     "postcss-cli": "^8.3.1",
     "preval.macro": "^5.0.0",
-    "react": "0.0.0-experimental-e5d06e34b",
-    "react-dom": "0.0.0-experimental-e5d06e34b",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "resolve": "^1.20.0",
     "typescript": "^4.2.3",
     "whatwg-fetch": "^3.6.2"
