@@ -6,6 +6,7 @@ import '@reach/tooltip/styles.css'
 
 import * as React from 'react'
 import ReactDOM from 'react-dom'
+import {createRoot} from 'react-dom/client'
 import {FaTools} from 'react-icons/fa'
 import {Tooltip} from '@reach/tooltip'
 import {Tabs, TabList, TabPanels, TabPanel, Tab} from '@reach/tabs'
@@ -228,7 +229,7 @@ function install() {
     devToolsRoot.id = 'dev-tools'
     document.body.appendChild(devToolsRoot)
   }
-  ReactDOM.render(<DevTools />, devToolsRoot)
+  createRoot(devToolsRoot).render(<DevTools />)
 }
 
 function ControlsPanel() {
