@@ -7,7 +7,7 @@
 /* eslint-disable */
 /* tslint:disable */
 
-const INTEGRITY_CHECKSUM = '795882c72c7304f6fa1d4a65a2418900'
+const INTEGRITY_CHECKSUM = '82ef9b96d8393b6da34527d1d6e19187'
 const bypassHeaderName = 'x-msw-bypass'
 const activeClientIds = new Set()
 
@@ -114,7 +114,7 @@ async function handleRequest(event, requestId) {
   // Send back the response clone for the "response:*" life-cycle events.
   // Ensure MSW is active and ready to handle the message, otherwise
   // this message will pend indefinitely.
-  if (activeClientIds.has(client.id)) {
+  if (client && activeClientIds.has(client.id)) {
     ;(async function () {
       const clonedResponse = response.clone()
       sendToClient(client, {
@@ -138,7 +138,7 @@ async function handleRequest(event, requestId) {
 }
 
 async function getResponse(event, client, requestId) {
-  const { request } = event
+  const {request} = event
   const requestClone = request.clone()
   const getOriginalResponse = () => fetch(requestClone)
 
@@ -207,7 +207,7 @@ async function getResponse(event, client, requestId) {
     }
 
     case 'NETWORK_ERROR': {
-      const { name, message } = clientMessage.payload
+      const {name, message} = clientMessage.payload
       const networkError = new Error(message)
       networkError.name = name
 
@@ -240,7 +240,7 @@ If you wish to mock an error response, please refer to this guide: https://mswjs
 }
 
 self.addEventListener('fetch', function (event) {
-  const { request } = event
+  const {request} = event
 
   // Bypass navigation requests.
   if (request.mode === 'navigate') {
